kits listo con nro de stock y falta actualizar las ventas ingresos y salidas

+++ app/Http/Controllers/Admin/ProductController.php
@@ -7,6 +7,8 @@ use App\Models\Category;
 use App\Models\Inventario;
 use App\Models\Detalleingreso;
 use App\Models\Detalleventa;
+use App\Models\Detallecotizacion;
+use App\Models\Detalleinventario;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ProductFormRequest;
@@ -64,7 +66,9 @@ class ProductController extends Controller
     {
         $lascategorias = Category::all()->where('status','=',0);
         $product = Product::findOrFail($product_id);
-        $micategoria =Category::all()->where('id','=',$product->category_id);
+        $micategoria =Category::all()
+        ->where('id','=',$product->category_id)
+        ->where('status','=',1);
         if($micategoria){
             $categories = $lascategorias->concat($micategoria);
         }
@@ -105,12 +109,13 @@ class ProductController extends Controller
     public function destroy(int $product_id)
     {
         $product = Product::find($product_id);
-
-        $inventario = Inventario::all()->where('product_id','=',$product_id); 
+        $inv = Inventario::all()->where('product_id','=',$product_id)->first(); 
+        $inventario = Detalleinventario::all()->where('product_id','=',$inv->id); 
         $ingreso = Detalleingreso::all()->where('product_id','=',$product_id); 
-        $venta = Detalleventa::all()->where('product_id','=',$product_id); 
+        $venta = Detalleventa::all()->where('product_id','=',$product_id);
+        $cotizacion = Detallecotizacion::all()->where('product_id','=',$product_id); 
 
-        if(count($inventario)==0 && count($ingreso)==0 && count($venta)==0){ 
+        if(count($inventario)==0 && count($ingreso)==0 && count($venta)==0  && count($cotizacion)==0){ 
             $product->delete();
             return redirect()->back()->with('message','Producto Eliminado');
             $this->dispatchBrowserEvent('close-modal');
